fix for type conformability
comparison of integer expressions of different signedness: 'int8_t'
{aka 'signed char'} and 'size_t' {aka 'unsigned int'}

+++ Adafruit_GFX.cpp
@@ -2664,7 +2664,7 @@ void GFXcanvas16::drawFastRawHLine(int16_t x, int16_t y, int16_t w,
                                    uint16_t color) {
   // x & y already in raw (rotation 0) coordinates, no need to transform.
   size_t buffer_index = y * WIDTH + x;
-  for (int16_t i = buffer_index; i < buffer_index + w; i++) {
+  for (size_t i = buffer_index; i < buffer_index + w; i++) {
     buffer[i] = color;
   }
 }
